All x(), y() and z() methods changed into ix(), iy() and iz()

+++ core/math/src/main/java/org/arakhne/afc/math/continous/object3d/Point3f.java
@@ -60,7 +60,7 @@ public class Point3f extends Tuple3f<Point3D> implements Point3D {
 	/**
 	 * @param tuple is the tuple to copy.
 	 */
-	public Point3f(float[] tuple) {
+	public Point3f(double[] tuple) {
 		super(tuple);
 	}
 
@@ -79,7 +79,7 @@ public class Point3f extends Tuple3f<Point3D> implements Point3D {
 	 * @param z
 	 */
 	public Point3f(float x, float y, float z) {
-		super(x,y,z);
+		super((double)x,(double)y,(double)z);
 	}
 
 	/**
@@ -88,7 +88,7 @@ public class Point3f extends Tuple3f<Point3D> implements Point3D {
 	 * @param z
 	 */
 	public Point3f(double x, double y, double z) {
-		super((float)x,(float)y,(float)z);
+		super(x,y,z);
 	}
 
 	/**
@@ -112,7 +112,7 @@ public class Point3f extends Tuple3f<Point3D> implements Point3D {
 	 */
 	@Override
 	public int distanceSquared(Point3D p1) {
-	      float dx, dy, dz;
+		 double dx, dy, dz;
 	      dx = this.x-p1.getX();  
 	      dy = this.y-p1.getY();
 	      dz = this.z-p1.getZ();
@@ -123,8 +123,8 @@ public class Point3f extends Tuple3f<Point3D> implements Point3D {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public float getDistanceSquared(Point3D p1) {
-	      float dx, dy, dz;
+	public double getDistanceSquared(Point3D p1) {
+		double dx, dy, dz;
 	      dx = this.x-p1.getX();  
 	      dy = this.y-p1.getY();
 	      dz = this.z-p1.getZ();
@@ -136,7 +136,7 @@ public class Point3f extends Tuple3f<Point3D> implements Point3D {
 	 */
 	@Override
 	public int distance(Point3D p1) {
-	      float  dx, dy, dz;
+		double  dx, dy, dz;
 	      dx = this.x-p1.getX();  
 	      dy = this.y-p1.getY();
 	      dz = this.z-p1.getZ();
@@ -147,12 +147,12 @@ public class Point3f extends Tuple3f<Point3D> implements Point3D {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public float getDistance(Point3D p1) {
-	      float  dx, dy, dz;
+	public double getDistance(Point3D p1) {
+		double  dx, dy, dz;
 	      dx = this.x-p1.getX();  
 	      dy = this.y-p1.getY();
 	      dz = this.z-p1.getZ();
-	      return (float)Math.sqrt(dx*dx+dy*dy+dz*dz);
+	      return (double)Math.sqrt(dx*dx+dy*dy+dz*dz);
 	}
 
 	/**
@@ -167,7 +167,7 @@ public class Point3f extends Tuple3f<Point3D> implements Point3D {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public float getDistanceL1(Point3D p1) {
+	public double getDistanceL1(Point3D p1) {
 	      return Math.abs(this.x-p1.getX()) + Math.abs(this.y-p1.getY() + Math.abs(this.z-p1.getZ()));
 	}
 
@@ -184,7 +184,7 @@ public class Point3f extends Tuple3f<Point3D> implements Point3D {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public float getDistanceLinf(Point3D p1) {
+	public double getDistanceLinf(Point3D p1) {
 	      return Math.max( Math.max( Math.abs(this.x-p1.getX()), Math.abs(this.y-p1.getY())),
 	    		  Math.abs(this.z-p1.getZ()));
 	}
@@ -218,7 +218,7 @@ public class Point3f extends Tuple3f<Point3D> implements Point3D {
 	}
 
 	@Override
-	public void scaleAdd(float s, Vector3D t1, Point3D t2) {
+	public void scaleAdd(double s, Vector3D t1, Point3D t2) {
 		this.x = s * t1.getX() + t2.getX();
 		this.y = s * t1.getY() + t2.getY();
 		this.z = s * t1.getZ() + t2.getZ();
@@ -232,7 +232,7 @@ public class Point3f extends Tuple3f<Point3D> implements Point3D {
 	}
 
 	@Override
-	public void scaleAdd(float s, Point3D t1, Vector3D t2) {
+	public void scaleAdd(double s, Point3D t1, Vector3D t2) {
 		this.x = s * t1.getX() + t2.getX();
 		this.y = s * t1.getY() + t2.getY();
 		this.z = s * t1.getZ() + t2.getZ();
@@ -246,7 +246,7 @@ public class Point3f extends Tuple3f<Point3D> implements Point3D {
 	}
 
 	@Override
-	public void scaleAdd(float s, Vector3D t1) {
+	public void scaleAdd(double s, Vector3D t1) {
 		this.x = s * this.x + t1.getX();
 		this.y = s * this.y + t1.getY();
 		this.z = s * this.z + t1.getZ();
