+++ core/math/src/main/java/org/arakhne/afc/math/continous/object3d/Tuple3fComparator.java
@@ -48,11 +48,11 @@ public class Tuple3fComparator implements Comparator<Tuple3f<?>> {
 		if (o1==o2) return 0;
 		if (o1==null) return Integer.MIN_VALUE;
 		if (o2==null) return Integer.MAX_VALUE;
-		int cmp = Float.compare(o1.getX(), o2.getX());
+		int cmp = Double.compare(o1.getX(), o2.getX());
 		if (cmp!=0) return cmp;
-		cmp = Float.compare(o1.getY(), o2.getY());
+		cmp = Double.compare(o1.getY(), o2.getY());
 		if (cmp!=0) return cmp;
-		return Float.compare(o1.getZ(), o2.getZ());
+		return Double.compare(o1.getZ(), o2.getZ());
 	}
 		
 }
