+++ core/math/src/main/java/org/arakhne/afc/math/continous/object3d/Vector3f.java
@@ -62,7 +62,7 @@ public class Vector3f extends Tuple3f<Vector3D> implements Vector3D {
 	/**
 	 * @param tuple is the tuple to copy.
 	 */
-	public Vector3f(float[] tuple) {
+	public Vector3f(double[] tuple) {
 		super(tuple);
 	}
 
@@ -81,7 +81,7 @@ public class Vector3f extends Tuple3f<Vector3D> implements Vector3D {
 	 * @param z
 	 */
 	public Vector3f(float x, float y, float z) {
-		super(x,y,z);
+		super((double)x,(double)y,(double)z);
 	}
 
 	/**
@@ -90,7 +90,7 @@ public class Vector3f extends Tuple3f<Vector3D> implements Vector3D {
 	 * @param z
 	 */
 	public Vector3f(double x, double y, double z) {
-		super((float)x,(float)y,(float)z);
+		super(x,y,z);
 	}
 
 	/**
@@ -113,18 +113,18 @@ public class Vector3f extends Tuple3f<Vector3D> implements Vector3D {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public float angle(Vector3D v1) {
+	public double angle(Vector3D v1) {
 		double vDot = dot(v1) / ( length()*v1.length() );
 		if( vDot < -1.) vDot = -1.;
 		if( vDot >  1.) vDot =  1.;
-		return((float) (Math.acos( vDot )));
+		return((double) (Math.acos( vDot )));
 	}
 
 	/**
 	 * {@inheritDoc}
 	 */
 	@Override
-	public float dot(Vector3D v1) {
+	public double dot(Vector3D v1) {
 		return (this.x*v1.getX() + this.y*v1.getY() + this.z*v1.getZ());
 	}
 
@@ -136,9 +136,9 @@ public class Vector3f extends Tuple3f<Vector3D> implements Vector3D {
 	 */
 	public final Vector3f mul(Matrix3f m) {
 		Vector3f r = new Vector3f();
-		r.x = (float) (this.getX() * m.m00 + this.getY() * m.m01 + this.getZ() * m.m02);
-		r.y = (float) (this.getX() * m.m10 + this.getY() * m.m11 + this.getZ() * m.m12);
-		r.z = (float) (this.getX() * m.m20 + this.getY() * m.m21 + this.getZ() * m.m22);
+		r.x = (double) (this.getX() * m.m00 + this.getY() * m.m01 + this.getZ() * m.m02);
+		r.y = (double) (this.getX() * m.m10 + this.getY() * m.m11 + this.getZ() * m.m12);
+		r.z = (double) (this.getX() * m.m20 + this.getY() * m.m21 + this.getZ() * m.m22);
 		return r;
 	}
 
@@ -154,33 +154,33 @@ public class Vector3f extends Tuple3f<Vector3D> implements Vector3D {
 
 	@Override
 	public Vector3D crossLeftHand(Vector3D v1) {
-		float x = v1.getY()*getZ() - v1.getZ()*getY();
-		float y = v1.getZ()*getX() - v1.getX()*getZ();
-		float z = v1.getX()*getY() - v1.getY()*getX();
+		double x = v1.getY()*getZ() - v1.getZ()*getY();
+		double y = v1.getZ()*getX() - v1.getX()*getZ();
+		double z = v1.getX()*getY() - v1.getY()*getX();
 		return new Vector3f(x,y,z);
 	}
 
 	@Override
 	public void crossLeftHand(Vector3D v1, Vector3D v2) {
-		float x = v2.getY()*v1.getZ() - v2.getZ()*v1.getY();
-		float y = v2.getZ()*v1.getX() - v2.getX()*v1.getZ();
-		float z = v2.getX()*v1.getY() - v2.getY()*v1.getX();
+		double x = v2.getY()*v1.getZ() - v2.getZ()*v1.getY();
+		double y = v2.getZ()*v1.getX() - v2.getX()*v1.getZ();
+		double z = v2.getX()*v1.getY() - v2.getY()*v1.getX();
 		set(x,y,z);
 	}
 
 	@Override
 	public Vector3D crossRightHand(Vector3D v1) {
-		float x = getY()*v1.getZ() - getZ()*v1.getY();
-		float y = getZ()*v1.getX() - getX()*v1.getZ();
-		float z = getX()*v1.getY() - getY()*v1.getX();
+		double x = getY()*v1.getZ() - getZ()*v1.getY();
+		double y = getZ()*v1.getX() - getX()*v1.getZ();
+		double z = getX()*v1.getY() - getY()*v1.getX();
 		return new Vector3f(x,y,z);
 	}
 
 	@Override
 	public void crossRightHand(Vector3D v1, Vector3D v2) {
-		float x = v1.getY()*v2.getZ() - v1.getZ()*v2.getY();
-		float y = v1.getZ()*v2.getX() - v1.getX()*v2.getZ();
-		float z = v1.getX()*v2.getY() - v1.getY()*v2.getX();
+		double x = v1.getY()*v2.getZ() - v1.getZ()*v2.getY();
+		double y = v1.getZ()*v2.getX() - v1.getX()*v2.getZ();
+		double z = v1.getX()*v2.getY() - v1.getY()*v2.getX();
 		set(x,y,z);
 	}
 
@@ -188,15 +188,15 @@ public class Vector3f extends Tuple3f<Vector3D> implements Vector3D {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public float length() {
-        return (float) Math.sqrt(this.x*this.x + this.y*this.y + this.z*this.z);
+	public double length() {
+        return (double) Math.sqrt(this.x*this.x + this.y*this.y + this.z*this.z);
 	}
 
 	/**
 	 * {@inheritDoc}
 	 */
 	@Override
-	public float lengthSquared() {
+	public double lengthSquared() {
         return (this.x*this.x + this.y*this.y + this.z*this.z);
 	}
 
@@ -205,7 +205,7 @@ public class Vector3f extends Tuple3f<Vector3D> implements Vector3D {
 	 */
 	@Override
 	public void normalize(Vector3D v1) {
-        float norm = 1f / v1.length();
+		double norm = 1f / v1.length();
         this.x = (int)(v1.getX()*norm);
         this.y = (int)(v1.getY()*norm);
         this.z = (int)(v1.getZ()*norm);
@@ -216,8 +216,8 @@ public class Vector3f extends Tuple3f<Vector3D> implements Vector3D {
 	 */
 	@Override
 	public void normalize() {
-        float norm;
-        norm = (float)(1./Math.sqrt(this.x*this.x + this.y*this.y + this.z*this.z));
+		double norm;
+        norm = (double)(1./Math.sqrt(this.x*this.x + this.y*this.y + this.z*this.z));
         this.x *= norm;
         this.y *= norm;
         this.z *= norm;
@@ -227,7 +227,7 @@ public class Vector3f extends Tuple3f<Vector3D> implements Vector3D {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public void turnVector(Vector3D axis, float angle) {
+	public void turnVector(Vector3D axis, double angle) {
 		Transform3D mat = new Transform3D();
 		mat.setRotation(new Quaternion(axis, angle));
 		mat.transform(this);
@@ -267,7 +267,7 @@ public class Vector3f extends Tuple3f<Vector3D> implements Vector3D {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public void scaleAdd(float s, Vector3D t1, Vector3D t2) {
+	public void scaleAdd(double s, Vector3D t1, Vector3D t2) {
 		this.x = s * t1.getX() + t2.getX();
 		this.y = s * t1.getY() + t2.getY();
 		this.z = s * t1.getZ() + t2.getZ();
@@ -287,7 +287,7 @@ public class Vector3f extends Tuple3f<Vector3D> implements Vector3D {
 	 * {@inheritDoc}
 	 */
 	@Override
-	public void scaleAdd(float s, Vector3D t1) {
+	public void scaleAdd(double s, Vector3D t1) {
 		this.x = s * this.x + t1.getX();
 		this.y = s * this.y + t1.getY();
 		this.z = s * this.z + t1.getZ();
